% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{coef,stantvafit-method}
\alias{coef,stantvafit-method}
\title{Model coefficients}
\usage{
\S4method{coef}{stantvafit}(object)
}
\arguments{
\item{object}{The StanTVA fit object.}
}
\value{
The model coefficients, grouped by random factors.
}
\description{
Returns the model coefficients (sum of fixed + random effects, grouped by random factor) for a StanTVA fit object.
}
\examples{
\donttest{
fixef <- coef(fit)
fixef
}
}
