% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{fixef,stantvafit-method}
\alias{fixef,stantvafit-method}
\title{Fixed effects}
\usage{
\S4method{fixef}{stantvafit}(object)
}
\arguments{
\item{object}{The StanTVA fit object.}
}
\value{
The fixed effects.
}
\description{
Returns the fixed effects for a StanTVA fit object.
}
\examples{
\donttest{
fixed_effects <- fixef(fit)
fixed_effects
}
}
