% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{flatten_matrix_col}
\alias{flatten_matrix_col}
\title{Flatten matrix columns}
\usage{
flatten_matrix_col(mat, collapse = ",", na_string = "")
}
\arguments{
\item{mat}{A character matrix}

\item{collapse}{String to use for collapsing elements (default: \code{,})}

\item{na_string}{String to use instead of \code{NA} values}
}
\value{
Character vector of concatenated rows
}
\description{
Collapse all rows of \code{mat} into single character strings, with columns separated by \code{collapse}. \code{NA}s can be replaced by \code{na_string}.
}
