% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{logLik,stantvafit-method}
\alias{logLik,stantvafit-method}
\title{Log-likelihood}
\usage{
\S4method{logLik}{stantvafit}(object)
}
\arguments{
\item{object}{The StanTVA fit.}
}
\value{
The pointwise log likelihood.
}
\description{
Returns the pointwise log-likelihood of a StanTVA fit.
}
\examples{
\donttest{
loglik <- logLik(model, data, params)
loglik
}
}
