% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{names,stantvafit-method}
\alias{names,stantvafit-method}
\title{Retrieve model parameter names}
\usage{
\S4method{names}{stantvafit}(x)
}
\arguments{
\item{x}{The StanTVA fit.}
}
\value{
The list of parameter names and aliases.
}
\description{
Returns the names of the fitted model parameters.
}
