% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{predict,stantvafit-method}
\alias{predict,stantvafit-method}
\title{Predict parameter values}
\usage{
\S4method{predict}{stantvafit}(object, newdata, variables = names(object@stanmodel@code@df))
}
\arguments{
\item{object}{The StanTVA fit object.}

\item{newdata}{The new data (leave empty to use fitted data).}

\item{variables}{The names of the parameters to predict.}
}
\value{
The predictions.
}
\description{
Returns the predictions for latent model parameters.
}
\examples{
\donttest{
p <- predict(fit, variables = c("C","K"))
colMeans(p$C)
}
}
