% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{show,stantvafit-method}
\alias{show,stantvafit-method}
\alias{print,stantvafit-method}
\title{Print StanTVA fit}
\usage{
\S4method{show}{stantvafit}(object)

\S4method{print}{stantvafit}(x, digits_summary = 2, ...)
}
\arguments{
\item{object}{The StanTVA fit object.}

\item{x}{The StanTVA fit object.}

\item{digits_summary}{The number of significant digits to display in posterior summaries.}

\item{...}{Currently not used.}
}
\value{
Returns \code{x}. Usually called for its side effects (printing to the console).
}
\description{
Prints a StanTVA fit object.
}
\section{Functions}{
\itemize{
\item \code{show(stantvafit)}: Alias

}}
\examples{
\donttest{
print(fit)
}
}
