% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{show,stantvafit-method}
\alias{show,stantvafit-method}
\title{Show StanTVA fit object}
\usage{
\S4method{show}{stantvafit}(object)
}
\arguments{
\item{object}{The StanTVA fit object.}
}
\description{
Show StanTVA fit object
}
\section{Methods (by class)}{
\itemize{
\item \code{show(stantvafit)}: Alias

}}
