% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{read_stantva_fit}
\alias{read_stantva_fit}
\title{Read StanTVA fit}
\usage{
read_stantva_fit(files)
}
\arguments{
\item{files}{The file names.}
}
\value{
The StanTVA fit object.
}
\description{
Reads a StanTVA fit object from one or more files. If multiple files are given, the fits will be combined into a single fit object (e.g., combining separately fitted chains).
}
\examples{
\donttest{
fit <- read_stantva_fit(c("chain1.rds", "chain2.rds"))
fit
}
}
