% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{read_tva_data}
\alias{read_tva_data}
\title{Read TVA data}
\usage{
read_tva_data(file, set = LETTERS, ...)
}
\arguments{
\item{file}{The file name.}

\item{set}{The set of items.}

\item{...}{Additional arguments passed to \code{\link[readr:read_table]{read_table()}}.}
}
\value{
A TVA data object, which inherits from \code{data.frame}.
}
\description{
Reads TVA data from a file.
}
\examples{
\donttest{
data <- read_tva_data("data.dat")
data
}
}
