% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{show,stantvacode-method}
\alias{show,stantvacode-method}
\title{Show StanTVA code}
\usage{
\S4method{show}{stantvacode}(object)
}
\arguments{
\item{object}{The StanTVA code object.}
}
\value{
Returns \code{object} invisibly but the function is usually only called for its side effects.
}
\description{
Display the content of the StanTVA code object in the console.
}
