% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{show,stantvamodel-method}
\alias{show,stantvamodel-method}
\title{Show StanTVA model}
\usage{
\S4method{show}{stantvamodel}(object)
}
\arguments{
\item{object}{The StanTVA model object.}
}
\value{
The printed object.
}
\description{
Prints a StanTVA model object.
}
\examples{
\donttest{
model <- stantva_model(locations = 4)
show(model)
}
}
