% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{stantva_model}
\alias{stantva_model}
\title{StanTVA model}
\usage{
stantva_model(..., stan_options = list())
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[RStanTVA:stantva_code]{stantva_code()}}.}

\item{stan_options}{The Stan options, passed to \code{\link[rstan:stan_model]{stan_model()}}}
}
\value{
The StanTVA model object.
}
\description{
Creates a StanTVA model object.
}
\examples{
\donttest{
model <- stantva_model(locations = 2, task = "pr")
model
}
}
