% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{stantva_path}
\alias{stantva_path}
\title{StanTVA path}
\usage{
stantva_path()
}
\value{
The path to the StanTVA directory.
}
\description{
Returns the path to the StanTVA directory.
}
\details{
This function is used internally by the \code{\link[RStanTVA:stantva_model]{stantva_model()}} method.
}
\examples{
path <- stantva_path()
path
}
