% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\docType{class}
\name{stantvacode-class}
\alias{stantvacode-class}
\title{StanTVA code class}
\description{
StanTVA code class
}
\section{Slots}{

\describe{
\item{\code{code}}{The generated Stan code.}

\item{\code{config}}{A list of model configuration parameters, as passed to \code{stantva_code()} or \code{stantva_model()}.}

\item{\code{include_path}}{The path to the StanTVA includes (usually identical to \code{stantva_path()}).}

\item{\code{df}}{The degrees of freedom of the model parameters.}

\item{\code{dim}}{The dimensions of the model parameters.}

\item{\code{version}}{The RStanTVA package version that was used to generate this model fit.}

\item{\code{priors}}{Priors for the model parameters.}
}}

