% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\docType{class}
\name{stantvafit-class}
\alias{stantvafit-class}
\title{StanTVA fit class}
\description{
StanTVA fit class
}
\section{Slots}{

\describe{
\item{\code{stanmodel}}{The StanTVA model object that was fitted to the data.}

\item{\code{data}}{The data to which the StanTVA model was fitted.}
}}

