% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\docType{class}
\name{stantvamodel-class}
\alias{stantvamodel-class}
\title{StanTVA model class}
\description{
StanTVA model class
}
\section{Slots}{

\describe{
\item{\code{code}}{The StanTVA code object that was used to compile this model.}
}}

