% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_plot_predicted_score}
\alias{tva_plot_predicted_score}
\title{Plot expected scores}
\usage{
tva_plot_predicted_score(
  nS,
  nD,
  tva_model,
  tva_fit,
  scores = 1:tva_model@code@config$locations,
  xlim = c(0, 200),
  n = 101
)
}
\arguments{
\item{nS}{Number of stimuli.}

\item{nD}{Number of distractors.}

\item{tva_model}{TVA model.}

\item{tva_fit}{Fitted parameters.}

\item{scores}{Scores to plot.}

\item{xlim}{Range of exposure durations.}

\item{n}{Number of grid points.}
}
\value{
A ggplot object.
}
\description{
Visualizes expected score as a function of exposure duration.
}
