% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_plot_score_densities}
\alias{tva_plot_score_densities}
\title{Plot score densities}
\usage{
tva_plot_score_densities(
  panels,
  tva_model,
  tva_fit,
  scores = 0:tva_model@code@config$locations,
  xlim = c(0, 200),
  n = 101
)
}
\arguments{
\item{tva_model}{TVA model.}

\item{tva_fit}{Fitted parameters.}

\item{scores}{Scores to plot.}

\item{xlim}{Range of exposure durations.}

\item{n}{Number of grid points.}

\item{nS}{Number of stimuli.}

\item{nD}{Number of distractors.}

\item{locations}{Number of locations.}
}
\value{
A ggplot object.
}
\description{
Visualizes score probability densities across exposure durations.
}
