% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_predict_score}
\alias{tva_predict_score}
\title{Predict expected score}
\usage{
tva_predict_score(
  S = matrix(1L, tva_model@code@config$locations, length(T)),
  D = matrix(0L, tva_model@code@config$locations, length(T)),
  T,
  scores = 1:tva_model@code@config$locations,
  tva_model,
  tva_fit
)
}
\arguments{
\item{S}{Integer matrix indicating stimulus presence.}

\item{D}{Integer matrix indicating distractor presence.}

\item{T}{Numeric vector of exposure durations.}

\item{tva_model}{TVA model object.}

\item{tva_fit}{Fitted TVA model parameters.}
}
\value{
Numeric vector of expected scores.
}
\description{
Computes the expected score based on TVA score probabilities.
}
