% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_predict_score_by_list}
\alias{tva_predict_score_by_list}
\title{Predict expected score}
\usage{
tva_predict_score_by_list(Tlist, tva_data, tva_model, tva_fit)
}
\arguments{
\item{Tlist}{A list of lists of exposure durations}

\item{tva_data}{TVA data for which to calculate the processing rates.}

\item{tva_model}{TVA model object.}

\item{tva_fit}{Fitted TVA model parameters.}
}
\value{
Predicted scores as a data frame
}
\description{
Compute the expected score for all combinations of \code{tva_data} and \code{Tlist}.
}
