% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_processing_rates}
\alias{tva_processing_rates}
\title{Compute TVA processing rates}
\usage{
tva_processing_rates(tva_data, tva_model, tva_fit)
}
\arguments{
\item{tva_data}{TVA data for which to calculate the processing rates.}

\item{tva_model}{TVA model object.}

\item{tva_fit}{Fitted TVA model parameters.}
}
\value{
A matrix of processing rates.
}
\description{
Computes processing rates \code{v} for given stimulus (\code{S}), distractor (\code{D}),
and exposure durations (\code{T}) using a fitted TVA model.
}
