% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tva_recovery_true_params}
\alias{tva_recovery_true_params}
\title{True parameters for TVA recovery study}
\format{
An object of class \code{list} of length 5.
}
\usage{
tva_recovery_true_params
}
\description{
True parameters for TVA recovery study
}
\keyword{datasets}
