% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{tva_report}
\alias{tva_report}
\title{Generate typical descriptive statistics for TVA reports}
\usage{
tva_report(data)
}
\arguments{
\item{data}{The TVA report data as a \code{data.frame}.}
}
\value{
The function returns a transmuted \code{data.frame}/\code{tibble} with columns \code{condition} (copied from \code{data}), \code{exposure} (copied from \code{data$T}), \code{n_items}, \code{n_targets}, \code{n_distractors}, and \code{score} (number of correctly reported items).
}
\description{
This function generates by-trial descriptive statistics, see `Value` below.
}
\examples{
\donttest{tva_report(tva_recovery)}
}
