% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_score_prob}
\alias{tva_score_prob}
\title{Compute score probabilities}
\usage{
tva_score_prob(
  tva_data,
  tva_model,
  tva_fit,
  score = 0:ncol(tva_data$S),
  log_p = FALSE
)
}
\arguments{
\item{tva_data}{TVA data for which to calculate the processing rates.}

\item{tva_model}{TVA model object.}

\item{tva_fit}{Fitted TVA model parameters.}

\item{score}{Integer vector of scores.}

\item{log_p}{Logical; return log-probabilities if TRUE.}
}
\value{
Matrix of probabilities (or log-probabilities).
}
\description{
Calculates the probability of observing specific scores.
}
