% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_simulate_response}
\alias{tva_simulate_response}
\title{Simulate TVA response}
\usage{
tva_simulate_response(
  tva_data,
  tva_model,
  tva_fit,
  seed = sample.int(.Machine$integer.max, 1)
)
}
\arguments{
\item{tva_data}{TVA data for which to calculate the processing rates.}

\item{tva_model}{TVA model object.}

\item{tva_fit}{Fitted TVA model parameters.}

\item{seed}{Random seed.}
}
\value{
Matrix of simulated responses.
}
\description{
Simulates trial-wise TVA responses.
}
