% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{write_stantva_fit}
\alias{write_stantva_fit}
\title{Write StanTVA fit}
\usage{
write_stantva_fit(fit, file, ...)
}
\arguments{
\item{fit}{The StanTVA fit object.}

\item{file}{The file name.}

\item{...}{Additional arguments passed to \code{\link[base:saveRDS]{saveRDS()}}.}
}
\value{
No return value, called for side effects.
}
\description{
Writes a StanTVA fit object to a file.
}
\examples{
\donttest{write_stantva_fit(fit, "fit.rds")}
}
