% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{write_tva_data}
\alias{write_tva_data}
\title{Write TVA data}
\usage{
write_tva_data(data, file, ...)
}
\arguments{
\item{data}{The TVA data object.}

\item{file}{The file name.}

\item{...}{Additional arguments passed to \code{\link[readr:write_tsv]{write_tsv()}}.}
}
\value{
No return value, called for side effects.
}
\description{
Writes TVA data to a file.
}
\examples{
\donttest{
data <- read_tva_data("data.dat")
write_tva_data(data, "data.dat")
}
}
