% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tva.R
\name{alias,stantvafit-method}
\alias{alias,stantvafit-method}
\title{Retrieve parameters aliases}
\usage{
\S4method{alias}{stantvafit}(object)
}
\arguments{
\item{object}{The StanTVA fit object.}
}
\value{
A character vector of parameter aliases.
}
\description{
Returns the StanTVA parameter aliases for the underlying RStan fit.
}
\examples{
\donttest{
al <- alias(fit)
al
}
}
