% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RStanTVA-package.R
\docType{package}
\name{RStanTVA-package}
\alias{RStanTVA}
\alias{RStanTVA-package}
\title{RStanTVA: TVA models in Stan using R and StanTVA}
\description{
RStanTVA is an R package containing a Stan implementation of the Theory of Visual Attention (TVA; Bundesen, 1990; \doi{10.1037/0033-295X.97.4.523}) and numerous convenience functions for generating, compiling, fitting, and analyzing (RStan)TVA models.
}
\author{
\strong{Maintainer}: Maximilian M. Rabe \email{maximilian.rabe@uni-potsdam.de} (\href{https://orcid.org/0000-0002-2556-5644}{ORCID})

Authors:
\itemize{
  \item Søren Kyllingsbæk \email{soeren.kyllingsbaek@psy.ku.dk} (\href{https://orcid.org/0000-0002-4100-437X}{ORCID})
}

}
\keyword{internal}
