% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_item_prob}
\alias{tva_item_prob}
\title{Compute item-level probabilities}
\usage{
tva_item_prob(
  tva_data,
  tva_model,
  tva_fit,
  item = seq_len(ncol(tva_data$S)),
  log_p = FALSE
)
}
\arguments{
\item{tva_model}{TVA model object.}

\item{tva_fit}{Fitted TVA model parameters.}

\item{item}{Items to evaluate.}

\item{log_p}{Logical; return log-probabilities if TRUE.}
}
\value{
Matrix of probabilities.
}
\description{
Computes probability of correctly reporting each item.
}
