% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_item_prob}
\alias{tva_item_prob}
\title{Compute item-level probabilities}
\usage{
tva_item_prob(
  S = matrix(1L, length(T), tva_model@code@config$locations),
  D = matrix(0L, length(T), tva_model@code@config$locations),
  T,
  tva_model,
  tva_fit,
  item = seq_len(ncol(S)),
  log_p = FALSE
)
}
\arguments{
\item{S}{Integer matrix indicating stimulus presence.}

\item{D}{Integer matrix indicating distractor presence.}

\item{T}{Numeric vector of exposure durations.}

\item{tva_model}{TVA model object.}

\item{tva_fit}{Fitted TVA model parameters.}

\item{item}{Items to evaluate.}

\item{log_p}{Logical; return log-probabilities if TRUE.}
}
\value{
Matrix of probabilities.
}
\description{
Computes probability of correctly reporting each item.
}
