% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_predict_score}
\alias{tva_predict_score}
\title{Predict expected score}
\usage{
tva_predict_score(
  tva_data,
  tva_model,
  tva_fit,
  scores = 1:tva_model@code@config$locations
)
}
\arguments{
\item{tva_model}{TVA model object.}

\item{tva_fit}{Fitted TVA model parameters.}
}
\value{
Numeric vector of expected scores.
}
\description{
Computes the expected score based on TVA score probabilities.
}
