% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_processing_rates}
\alias{tva_processing_rates}
\title{Compute TVA processing rates}
\usage{
tva_processing_rates(
  S = matrix(1L, tva_model@code@config$locations, length(T)),
  D = matrix(0L, tva_model@code@config$locations, length(T)),
  T,
  tva_model,
  tva_fit
)
}
\arguments{
\item{S}{Integer matrix indicating stimulus presence.}

\item{D}{Integer matrix indicating distractor presence.}

\item{T}{Numeric vector of exposure durations.}

\item{tva_model}{TVA model object.}

\item{tva_fit}{Fitted TVA model parameters.}
}
\value{
A matrix of processing rates.
}
\description{
Computes processing rates \code{v} for given stimulus (\code{S}), distractor (\code{D}),
and exposure durations (\code{T}) using a fitted TVA model.
}
