% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tva_recovery}
\alias{tva_recovery}
\title{True parameters for TVA recovery study}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 25600 rows and 9 columns.
}
\usage{
tva_recovery
}
\description{
...
}
\keyword{datasets}
