% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_response_prob}
\alias{tva_response_prob}
\title{Compute response probability}
\usage{
tva_response_prob(
  R,
  S = matrix(1L, tva_model@code@config$locations, length(T)),
  D = matrix(0L, tva_model@code@config$locations, length(T)),
  T,
  tva_model,
  tva_fit,
  log_p = FALSE
)
}
\arguments{
\item{R}{Integer matrix of responses.}

\item{S}{Integer matrix indicating stimulus presence.}

\item{D}{Integer matrix indicating distractor presence.}

\item{T}{Numeric vector of exposure durations.}

\item{tva_model}{TVA model object.}

\item{tva_fit}{Fitted TVA model parameters.}

\item{log_p}{Logical; return log-probabilities if TRUE.}
}
\value{
Numeric vector of probabilities.
}
\description{
Computes the probability of a response pattern \code{R}.
}
