% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_response_prob}
\alias{tva_response_prob}
\title{Compute response probability}
\usage{
tva_response_prob(tva_data, tva_model, tva_fit, log_p = FALSE)
}
\arguments{
\item{tva_model}{TVA model object.}

\item{tva_fit}{Fitted TVA model parameters.}

\item{log_p}{Logical; return log-probabilities if TRUE.}

\item{R}{Integer matrix of responses.}
}
\value{
Numeric vector of probabilities.
}
\description{
Computes the probability of a response pattern \code{R}.
}
