% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_score_prob}
\alias{tva_score_prob}
\title{Compute score probabilities}
\usage{
tva_score_prob(
  S = matrix(1L, length(T), tva_model@code@config$locations),
  D = matrix(0L, length(T), tva_model@code@config$locations),
  T,
  tva_model,
  tva_fit,
  score = 0:ncol(S),
  log_p = FALSE
)
}
\arguments{
\item{S}{Integer matrix indicating stimulus presence.}

\item{D}{Integer matrix indicating distractor presence.}

\item{T}{Numeric vector of exposure durations.}

\item{tva_model}{TVA model object.}

\item{tva_fit}{Fitted TVA model parameters.}

\item{score}{Integer vector of scores.}

\item{log_p}{Logical; return log-probabilities if TRUE.}
}
\value{
Matrix of probabilities (or log-probabilities).
}
\description{
Calculates the probability of observing specific scores.
}
