% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_simulate_score}
\alias{tva_simulate_score}
\title{Simulate TVA score}
\usage{
tva_simulate_score(
  S = matrix(1L, tva_model@code@config$locations, length(T)),
  D = matrix(0L, tva_model@code@config$locations, length(T)),
  T,
  tva_model,
  tva_fit,
  seed = sample.int(.Machine$integer.max, 1)
)
}
\arguments{
\item{S}{Integer matrix indicating stimulus presence.}

\item{D}{Integer matrix indicating distractor presence.}

\item{T}{Numeric vector of exposure durations.}

\item{tva_model}{TVA model object.}

\item{tva_fit}{Fitted TVA model parameters.}

\item{seed}{Random seed.}
}
\value{
Numeric vector of simulated scores.
}
\description{
Simulates scores by summing simulated responses.
}
