% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{tva_simulate_score}
\alias{tva_simulate_score}
\title{Simulate TVA score}
\usage{
tva_simulate_score(
  tva_data,
  tva_model,
  tva_fit,
  seed = sample.int(.Machine$integer.max, 1)
)
}
\arguments{
\item{tva_model}{TVA model object.}

\item{tva_fit}{Fitted TVA model parameters.}

\item{seed}{Random seed.}
}
\value{
Numeric vector of simulated scores.
}
\description{
Simulates scores by summing simulated responses.
}
